+++ src/components.d.ts
@@ -157,10 +157,6 @@ export namespace Components {
     * URL-friendly slug (e.g. `"jawsdb-mysql"`)
     */
     'productLabel': string;
-    /**
-    * _(Deprecated)_ Look up product logo from resource
-    */
-    'resourceLabel'?: string;
   }
   interface ManifoldDataProductName {
     /**
@@ -592,14 +588,6 @@ export namespace Components {
     'options': Option[];
     'required'?: boolean;
   }
-  interface ManifoldServiceCard {
-    'hideUntilReady'?: boolean;
-    'isFeatured'?: boolean;
-    'preserveEvent'?: boolean;
-    'product'?: ProductCardQuery['product'];
-    'productLabel'?: string;
-    'productLinkFormat'?: string;
-  }
   interface ManifoldSkeletonImg {}
   interface ManifoldSkeletonText {}
   interface ManifoldTemplateCard {
@@ -954,12 +942,6 @@ declare global {
     new (): HTMLManifoldSelectElement;
   };
 
-  interface HTMLManifoldServiceCardElement extends Components.ManifoldServiceCard, HTMLStencilElement {}
-  var HTMLManifoldServiceCardElement: {
-    prototype: HTMLManifoldServiceCardElement;
-    new (): HTMLManifoldServiceCardElement;
-  };
-
   interface HTMLManifoldSkeletonImgElement extends Components.ManifoldSkeletonImg, HTMLStencilElement {}
   var HTMLManifoldSkeletonImgElement: {
     prototype: HTMLManifoldSkeletonImgElement;
@@ -1049,7 +1031,6 @@ declare global {
     'manifold-resource-status': HTMLManifoldResourceStatusElement;
     'manifold-resource-status-view': HTMLManifoldResourceStatusViewElement;
     'manifold-select': HTMLManifoldSelectElement;
-    'manifold-service-card': HTMLManifoldServiceCardElement;
     'manifold-skeleton-img': HTMLManifoldSkeletonImgElement;
     'manifold-skeleton-text': HTMLManifoldSkeletonTextElement;
     'manifold-template-card': HTMLManifoldTemplateCardElement;
@@ -1181,10 +1162,6 @@ declare namespace LocalJSX {
     * URL-friendly slug (e.g. `"jawsdb-mysql"`)
     */
     'productLabel'?: string;
-    /**
-    * _(Deprecated)_ Look up product logo from resource
-    */
-    'resourceLabel'?: string;
   }
   interface ManifoldDataProductName {
     /**
@@ -1641,14 +1618,6 @@ declare namespace LocalJSX {
     'options'?: Option[];
     'required'?: boolean;
   }
-  interface ManifoldServiceCard {
-    'hideUntilReady'?: boolean;
-    'isFeatured'?: boolean;
-    'preserveEvent'?: boolean;
-    'product'?: ProductCardQuery['product'];
-    'productLabel'?: string;
-    'productLinkFormat'?: string;
-  }
   interface ManifoldSkeletonImg {}
   interface ManifoldSkeletonText {}
   interface ManifoldTemplateCard {
@@ -1737,7 +1706,6 @@ declare namespace LocalJSX {
     'manifold-resource-status': ManifoldResourceStatus;
     'manifold-resource-status-view': ManifoldResourceStatusView;
     'manifold-select': ManifoldSelect;
-    'manifold-service-card': ManifoldServiceCard;
     'manifold-skeleton-img': ManifoldSkeletonImg;
     'manifold-skeleton-text': ManifoldSkeletonText;
     'manifold-template-card': ManifoldTemplateCard;
@@ -1806,7 +1774,6 @@ declare module "@stencil/core" {
       'manifold-resource-status': LocalJSX.ManifoldResourceStatus & JSXBase.HTMLAttributes<HTMLManifoldResourceStatusElement>;
       'manifold-resource-status-view': LocalJSX.ManifoldResourceStatusView & JSXBase.HTMLAttributes<HTMLManifoldResourceStatusViewElement>;
       'manifold-select': LocalJSX.ManifoldSelect & JSXBase.HTMLAttributes<HTMLManifoldSelectElement>;
-      'manifold-service-card': LocalJSX.ManifoldServiceCard & JSXBase.HTMLAttributes<HTMLManifoldServiceCardElement>;
       'manifold-skeleton-img': LocalJSX.ManifoldSkeletonImg & JSXBase.HTMLAttributes<HTMLManifoldSkeletonImgElement>;
       'manifold-skeleton-text': LocalJSX.ManifoldSkeletonText & JSXBase.HTMLAttributes<HTMLManifoldSkeletonTextElement>;
       'manifold-template-card': LocalJSX.ManifoldTemplateCard & JSXBase.HTMLAttributes<HTMLManifoldTemplateCardElement>;