+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0]
+
+### Removed
+- `resourceLabel` prop on `<manifold-data-product-logo>` component (previously deprecated in favor of `<manifold-data-resource-logo>`).
+- `manifold-service-card` component (previously deprecated in favor of `manifold-product-card`).
+
 ## [0.8.0]
 
 ### Added